Added Salsify Frame class
A frame manages several fragments. It also allows both ways of
working with it. Firstly it allows you to split up huge amounts of
data into several fragments to prepare them for sending. Secondly
it allows you to add fragments to an incomplete frame. For example
while receiving. In that case it will also check if certain conditions
are met.