Bugfixing. Images are now displayed correct
Receiver did not read actual fragment size. It instead read 1004 bytes
which is not the size of a salsify fragment. The salsify frame had
a mistake where it would ignore the last bytes of each fragment while
copying the sub arrays.

Decreased speed of sending frames and  added a couple of log messages.